requirements.txt
requests
beautifulsoup4
numpy
pandas
python-dotenv
langchain
openai
pinecone-client
transformers
pypdf
tiktoken